Allow custom degree in random_element of polynomial quotient ring
Currently, the `random_element` function of the polynomial quotient ring passes
all arguments through to the base ring. However, the random_element function
itself passes a default degree to the base ring, based on the degree of the
quotient ring. So, overwriting the degree in kwargs gives a TypeError.
Therefore, detect the `degree` argument manually, such that the user can
overrride the default degree.