build/pkgs/debugpy: skip when the system ipykernel is being used
The debugpy package is used only by ipykernel. While it is listed
unconditionally as a dependency, it is technically optional. Gentoo's
ipykernel, for example, does not depend on it; and in fact, Gentoo's
debugpy package is being removed (it is a pain to maintain).

This commit un-requires debugpy when the system ipykernel is being
used: the dependencies of the system package are assumed to be
satisfied.