Workaround for multiprocessing shared memory limits
This affects musl libc and macos, see:
https://github.com/sagemath/sage/pull/36741#issuecomment-2027357414

We are working on a better fix to avoid using multiprocessing in
sage.features, but this will do meanwhile.

There is agreement that disabling the lock is completely harmless, since
the counter is not really used for anything; disabling the lock fixes
the issues that we had on 10.4.beta0.