build/pkgs/glpk: drop GLPK patch that was rejected by upstream.
SageMath has been carrying a patch to GLPK's error handling that was
rejected as unsupported by upstream. The GLPK backend has been updated
to obviate this patch (unless the user ignores a warning and changes
the default solver...), so we can finally drop it and bring SageMath's
GLPK package to parity with the various distributions.