Fixup changes to reestablish #36676
The chain of immediate ancestors of this commit was created by
cherry-picking the relevant commits that comprise #36676. Replaying all
these commits (and the necessary conflict resolution) still leaves a
non-empty diff when comparing 10.4.beta2 to the previous #36676 and
10.4.beta3 to that cherry-picked branch. This commit makes sure that the
diff is trivial. The changes introduced here were likely introduced in
conflict resolution when merging develop into #36676. I did not replay
these conflict resolutions, so I do this here manually.