src/sage/geometry/cone*.py: drop critical_angles() method
Recent research has shown that (in many cases) the maximal angle does
not change too much if the cones are perturbed. This creates even more
of a safety net when computing the maximal angle, because falling back
to inexact arithmetic "perturbs" the cones.

The critical angle function, however, is not lower-semicontinuous as a
set-valued map. I don't like having a "debug" parameter that tells you
you're getting wrong results, and a user interface that requires you
to understand the subtle differences between upper- and
lower-semicontinuity for set-valued maps is not awesome. I think it's
best if we leave this out for now.