Make # abs tol compare over the complex numbers
For calculations over complex numbers that generate numeric noise, one
tends to create small but non-zero imaginary parts. This PR updates
the "# abs tol" tolerance setting to work over the complex numbers, as
the "abs" suggests complex numbers. The real and imaginary parts are
compared separately.

The ordinary "# tol" and "# rel tol" are left as is.

Fixes #36631