src/sage/features/giac.py: add new feature for the giac program
In preparation for adding a --disable-giac option, we add a new
feature that detects the presence of the "giac" executable. We already
have a feature for sage.libs.giac, but that only guards the libgiac
interface; we still have code that runs "giac" behind pexpect. This
will allow us to skip those tests when giac is not installed.