src/sage/symbolic/integration: make libgiac integration optional
We turn the libgiac integrator into a no-op if libgiac isn't
available. That way, in the default integration routines, it is
essentially skipped. If it were tried last, we could literally skip
it; but currently it lives between maxima and sympy and cannot be
rearanged without messing up some doctests.