src/sage/functions/orthogonal_polys.py: avoid global state in doctests
We have an mpmath test in this file that sets the global precision
(decimal places) and "pretty" variable for all of mpmath. Instead it
is safer to use a context manager for the precision, and to call
print(foo) to obtain the string representation of foo. This fixes
the following two test failures:

  Failed example:
    gegenbauer_mp(-7,0.5,0.3)
  Expected:
    0.1291811875
  Got:
    mpf('0.1291811874999999999999999952')

and

  Failed example:
    gegenbauer_mp(2+3j, -0.75, -1000j)
  Expected:
    (-5038991.358609026523401901 + 9414549.285447104177860806j)
  Got:
    mpc(real='-5038991.358609026523401901035', imag='9414549.285447104177860806274')