src/sage/libs/mpmath/ext_main.pyx: disable pretty printing for a test
One test in this file sets mp.pretty=True but is failing for me, most
likely due to the shared state inherent to mp.pretty. Since this test
is inside a TESTS block and not visible to end users, the simplest way
to fix it is to delete the mp.pretty=True and to look for the ugly
output.