src/sage/libs/mpmath/ext_main.pyx: fix two repr tests with clone()
There are two tests in this file that are attempting to test the
repr() of an mpmath number with mp.pretty=True set globally. The
shared state is causing test failures, but we can't simply leave it
off, because that's the point of the test. Instead we can clone()
the "mp" object to obtain a new one (whose state is not so global)
and then work with that instead.