src/sage/doctest/util.py: test _proc_stat_cpu_seconds() error paths
Now that _proc_stat_cpu_seconds() takes a path instead of a PID, we
add two new tests for error conditions. The first checks for too few
fields in the stat file, and the second checks for non-float
fields. Both use a temporary file whose path can now be passed in.