src/sage/doctest/forker.py: re-raise SignalErrors
If SystemExit occurs during a doctest, we re-raise it. The same thing
should be done for SignalError (from cysignals). Signals can occur at
any time, and the doctest runner shouldn't eat them.

In particular, this can lead to a Heisenbug where a signal will
terminate a doctest example before the timer has a chance to annotate
the it with the elapsed walltime and cputime. In that case, we get
an "impossible" AttributeError later on when we try to compute the
total walltime used by the example.