src/sage/parallel/decorate.py: disable a stress-dependent test
One test in this file sets the timeout to 100ms and expects an easy
example to complete within that amount of time. It can still fail,
though, if the machine is under load:

  File "src/sage/parallel/decorate.py", line 519, in
  sage.parallel.decorate.fork
    Failed example:
      g(5, m=5)
    Expected:
      8
  Got:
      Killing subprocess 62130 with input ((5,), {'m': 5}) which
      took too long
      8

This example isn't really the purpose of the doctest -- its intent is
to show that there are hard examples that will NOT complete in
100ms. (That is much easier to do.) All things considered, we are
better off just dropping this one flaky example.