src/sage/graphs/generic_graph.py: work around doctest hang
One doctest in this file is "hanging" on ARM64 and RISC-V as GLPK
tries courageously to solve a MIP. A tweak to the solver options
allows this problem to be solved on those two architectures without
affecting any others. This is unlikely to solve the general problem,
but it may buy us some time.

Closes https://github.com/sagemath/sage/issues/34575
Closes https://github.com/sagemath/sage/issues/38831