src/sage/parallel/map_reduce.py: fixup start_workers() test case
In commit a1818838 I made this test wait a bit longer (to be more
robust under load) and added a "long time" to the corresponding
start_workers() line. But, further down, there's another test that
calls finish(), which needs the workers to be started.  Without a
"long time" on the finish(), it can fail sans --long.

To fix it, we combine the two tests that call start_workers() and
finish(), and then mark the whole thing "long time". The finish() is
supposed to delete the workers created in setup_workers() as well, so
there's not much point in trying to make the "long time" more
fine-grained than that.