src/sage/features/info.py: new feature for GNU Info
Before we can replace our hand-rolled parser for Singular's info file,
we need to be able to detect the "info" program from the GNU Info
package. The goal is to politely disable access to Singular's info if
GNU Info is unavailable. This avoids a hard dependency on GNU Info in
the sage library.

We require v7 or later because a bugfix in Texinfo v7.0.0 corrects the
exit code from "info" when the requested node is not found. This allows
us to separate the "expected failures" from the truly unexpected ones.