src/sage/env.py: don't pass if mismatched SAGE_ROOT is None
In this file we test that SAGE_ROOT in a subprocess is the same as
SAGE_ROOT in the parent. There is a special case for when SAGE_ROOT is
None, but that special case can pass if only one of the SAGE_ROOT
variables is None and the other is not -- contrary to the intent of
the test.

Here we extend the special case to ensure that both SAGE_ROOT
variables are None if one of them is.

Thanks to Gonzalo Tornaría for the suggestion.