src/sage/misc/sageinspect.py: raise exceptions in _sage_getdoc_unformatted
Our _sage_getdoc_unformatted() catches all exceptions and turns them
into the empty string. The side effects of this are pretty bad: this
function is used to generate the help? strings in the sage REPL, and
if something goes wrong, the user will get a confusing blank docstring
rather than the error.

The change was introduced in issue 19671 as a workaround for a runtime
error raised inside flask. But we are no longer using flask, and
hopefully no other parts of the Sage library raise an exception during
the documentation build. We are about to find out!