src/sage/algebras/letterplace: define singular functions locally
These modules use singular_function() from sage.libs.singular.function
to create top-level aliases for a few singular functions. The docbuild
however tries to build documentation for these functions, and it can
raise an error if GNU info is not installed. There's no good reason to
define them globally in the first place (they are only used locally),
so an easy fix is to move the imports and variables local.