src/sage/groups/libgap_mixin.py: hide info warning for isomorphism checks
GAP emits an "InfoWarning" message when it is checking for isomorphism
between two groups that are not known to be finite. This message is
only shown once, however, because after it checks for finiteness, it
knows. We have to keep this in mind during tests. For example,

  File "src/sage/groups/finitely_presented_named.py", line 485, in
  sage.groups.finitely_presented_named.AlternatingPresentation
  Failed example:
      A1.is_isomorphic(A2), A1.order()
  Expected:
      (True, 1)
  Got:
      #I  Forcing finiteness test
      (True, 1)

These warnings are "normal" and there's nothing for the user to do.
This commit hides them for the duration of the isomorphism check.

Fixes https://github.com/sagemath/sage/issues/38889