Correctly print import errors in `sage.all`
When there is an error during importing `sage.all`, then it is just hidden but sage's are no longer available. This is very hard to debug and misleading during development. For this reason the fallback to `repl.all` is removed.

Furthermore, a ruff exception for "unused imports" is added for all files, since these files always contains many unused imports.