src/sage/misc/table.py: fix the right border on table cells
It looks like during the conversion to unicode, the right borders of
table cells were overlooked. Instead of a nice unicode character, they
still have an ASCII pipe. For example,

  sage: table([1],frame=True)
  ┌───┐
  │ 1 |
  └───┘

This commit changes it to a unicode thingy, just like the left
borders,

  sage: table([1],frame=True)
  ┌───┐
  │ 1 │
  └───┘

and updates all of the tests in sage.misc.table.