PKG_CONFIG must be empty if not found
Followup to https://github.com/sagemath/sage/pull/38954, where I
followed the pkgconf documentation to set PKG_CONFIG=false if it is
not found. But Sage uses "test -z $PKG_CONFIG", so it must instead be
set to empty.

This causes pkgconf not to be built if pkgconf/pkg-config is not
found. Which is basically only macOS.