Supress ipython_genutils
This should not be a dependency of anything for the last ~7 years.
If it is; it is a bug in the package.

I think that sage is one of the last pieces of software that mention it
And might be the reason we still recieve request to remove nose as a
testing framework to package it in opensuse ?