Move bitness handling in doctest to the output checker
Currently, the bit marker `# 32/64-bit` is handled after parsing the doctests by modifying the doctest source. This handling is now moved to the output checker, which checks that the result of a doctest coincides with the expected value.

This change is made for compatibility with pytest, which itself handles the creation of doctests and thus doesn't invoke the `sources` handler (but allows to change the output checker).