Rename module-global test methods
Pytest automatically identifies and runs all methods prefixed with `test_`, following the standard convention for denoting tests. In the Sage codebase, there are numerous methods with the `test_` prefix. While many of these are indeed intended for testing purposes, they do not conform to Pytest's requirements for valid tests. To address this, we are temporarily renaming these methods, with the goal of eventually converting them into proper Pytest tests in the future.