src/sage/misc/latex.py: default the "engine" option to None
This delays probing the default engine (which tries to compile some
LaTeX code in a temporary directory) when you merely wish to query
e.g. the matrix delimiters. A later commit will re-insert the probe
at a later point.

The doctest for the options dict was updated to print a sorted list of
items, so now its output is no longer random.