src/sage/misc/latex.py: invoke default_engine() as late as possible
When we try to find the current LaTeX engine, the preferences dict
will now have __options["engine"] == None to indicate that we should
fall back to the default engine. In each case we check for None, and
invoke default_engine() only if necessary.

This allows us to query the options dict without triggering the
expensive is_functional() method for the LaTeX features.