Do not recommend to set MAKE=make -jX
Setting MAKE=make -jX can lead to an exponential growth of build
processes (as noted during SD128.) It's better to rely on make's
jobserver that is available on macOS and all relevant Linux
distributions.

We also recommend a sane default for MAKEFLAGS that people can just copy
& paste. Unfortunately, a lot of non-developers of SageMath still have
to build from source and they appreciate not having to think about these
things and just being able to copy a good default value.