move invariant bilinear form to class method
The computation of G-invariant symmetric bilinear forms only depends on the group and representation. It is a general computation. Since there doesn't seem to be a generic class for finite group representations, for now we leave this in the symmetric group representation generic class. Now it can be tested, optimized, and possibly moved in the future.

use self._ring and self._partition for self._specht

to define self._specht in SymmetricGroupRepresentations_generic_class, _partition defined in init along with _ring, so we can call those via self.

define self._specht to compute rho

self._spect is only defined for unitary representations

self._specht is defined within UnitaryRepresentation. here, we should be able to just get the representation matrix directly, and then we need the dimension. this can be computed as the trace of the identity usually, but since we're over a finite field, we should take the number of columns.

better comments

include 'symmetric'

not all invariant bilinear forms are symmetric. these are, so we really need to specify that.

add whitespace, check for zero dim'l nullspace

update doc examples, remove whitespace

compute U

only sum over generators

if a bilinear form is invariant under all generators, it is invariant for the entire group.

replace "the" with "an"

the solution is a vector space of some dimension, so the returned form is not the unique solution. even in the one dimensional case, it can be a scalar multiple.