move conj_sqrt to standalone method
conj_sqrt is a useful and common operation used for finite fields. it should be promoted to a standalone method. we should figure out an equivalence for the condition of the exponent being divisible by q+1. I originally thought that it was equivalent to being in the base field GF(q) of GF(q**2), but that doesn't hold for fields GF(q**r) where r is odd.