move to `SpechtRepresentation`
if we move this up to `SymmetricGroupRepresentation_generic_class`, then we have to be careful about calling `representation_matrix`. For instance, if we create a `UnitaryRepresentation`, then trying to compute an `invariant_symmetric_bilinear_form` will call `representation_matrix`, which will then call `invariant_symmetric_bilinear_form`. In other words, `representation_matrix` needs to be from a Specht module. It makes more sense to move this code to `SpechtRepresentation` because for these invariant forms, we are always working with respect to a representation. It's probably possible to genrealize this beyond SymmetricGroupRepresentations, or use GAP, but for now this makes the most sense. We still create `self._specht` in `UnitaryRepresentation`, now can just call `invariant_symmetric_bilinear_form()`