Improve NotImplementedError for sqrt(Mod(a,n), all=True)
When sqrt(Mod(a,n), all=True) is called for a non-square 'a' with the default extend=True, the resulting NotImplementedError was very generic.

Improved the error message to suggest trying extend=False to find roots only within the base ring Zmod(n), as discussed in the issue.

Included a doctest .