convert between two implementations of fin-dim algebras
This provides functionality for converting between two implementations of
finite-dimensional algebras:
- the modern one, implemented as MagmaticAlgebras.WithBasis.FiniteDimensional in
  sage.categories.magmatic_algebras (in the magmatic case) and as
  FiniteDimensionalAlgebrasWithBasis in
  sage.categories.finite_dimensional_algebras_with_basis
  (in the assocative case),
and
- the old one, implemented as FiniteDimensionalAlgebra in
  sage.algebras.finite_dimensional_algebras.finite_dimensional_algebra.
To be more precise, the latter now automatically inherits from the
former if the optional parameters assume_associative and
assume_unital are set to True, whereas the former can be
converted into the latter via the to_finite_dimensional_algebra method.
This way, the methods of one (e.g. ``center_basis`` on the modern
one or ``is_unitary`` on the old one) can be used on the other.

Fixing various other issues in both classes as well.