Disable most hashing for fraction field elements
previously hashing of fraction field elements relied on .reduce() to
produce a normal form of an element. This is however not the case. (And
also does not seem to be possible in general.)

Instead, we rather produce correct output here instead of just making
things fail in weird ways. This comes with a performance cost of course
as can be seen in the strange outputs of the hashes of function field
elements.