get_embedding: remove vertices that do not exist
An embedding is a Python dictionary d where d[v] is the list of neighbors
of vertex v. When vertices get deleted, the embedding is not updated
immediately so get_embedding() needs to remove all vertices in d[v] that
no longer exist. This commit fixes #40054 where not enough vertices in
d[v] were removed.