fix(Mpolynomial_libsingular): support division over transcendental extension
Currently, dividing a Singular backed multivariate polynomial by a fractional
field element(of the same type) is unsupported and triggers a conversion error
on the Singular end indicating the denominator is nonconstant. This commit
extracts the commit from Singular trunk which adds support for the use case[^1]
to a patch to be applied on top of release 4.4.1. We force an in-source build
of Singular if the system Singular package does not contain the patch.

[^1]: https://github.com/Singular/Singular/commit/0a4bc0e2888e660d1a5229a7b4842539a46c912a