src/sage/geometry/cone.py: add linear_isomorphisms() method
With the ultimate goal of reimplementing isomorphism testing for
convex cones, we add a new method to generate said isomorphisms. If
any are generated... then the cones are isomorphic.

This is based on the well-known decomposition of a closed convex cone
in to a sum of three convex cones:

  1. Its lineal part (the largest subspace contained in the cone)
  2. Its non-lineal part (the orthogonal complement to the lineal
     part)
  3. A trivial cone in a subspace orthogonal to the original cone (in
     the case where, say, the cone is 2d in a 3d space)

Generating isomorphisms for (1) and (3) are easy, because all we need
are vector space isomorphisms. A necessary condition for isomorphism
between cones that are both strictly convex is that extreme rays are
sent to extreme rays. And for a solid cone, there are enough extreme
rays to make any such map invertible. The cone in (2) satisfies both
of these conditions, so by decomposing two cones and matching up their
components, we are able to generate "all" ismorphisms between them (up
to some hand waving that has been documented in the ALGORITHM block.)