src/sage/geometry/cone.py: add is_linearly_isomorphic() method
This is a new alternative to the is_isomorphic() method for convex
cones. The existing method was written with toric geometry in mind,
and is neither documented nor intended to capture the notion of
equivalence under invertible rational transformations.

The new method was implemented "from scratch," avoiding the fan
machinery, and works for all convex cones, not just strictly convex
ones. It should also be immune to the symmetry bug in Github issue
37957.