src/sage/geometry/cone.py: synchronize equivalence documentation
In the documentation of is_equivalent() and is_isomorphic(), there is
a note that mentions "three notions of equivalence" for convex cones.
Well, with the addition of is_linearly_isomorphic(), now there are
four. We update the existing note in both places, and then copy/paste
it into the docs for is_linearly_isomorphic() as well.

We also add SEEALSO blocks for is_equivalent() and is_isomorphic() to
match the one in is_linearly_isomorphic().