src/sage/geometry/cone.py: faster containment for the trivial cone
When the cone K has no rays, the test "x in K" can be done quickly by
checking if x is zero. This can be a significant improvement if the
lattice is large, and risks wasting only as much time as it takes to
compare an integer to zero (i.e. nothing compared to how long the rest
of the containment test is going to take).