src/sage/geometry/cone.py: avoid two Cone() calls in isomorphism gen
When constructing linear_isomorphisms() between two cones, at one
point we call Cone() on two subsets of their respective generators to
check that the results have the same number of extreme rays. We can
however do this ourselves when we construct the list of rays that we
pass to Cone, by simply removing the zeros.

The remaining rays are guaranteed to be independent: we have projected
out the component in the lineality space; if the resulting rays are
conically dependent, then the original rays are too.