src/sage/geometry/cone.py: eliminate the last Cone() in linear_isomorphisms()
Rather than explicitly checking that the two Cone()s are equivalent
under a linear map, we can check that two sets of extreme rays get
mapped to each other. We already know that the map is invertible,
so this is sufficient and should be a bit faster than invoking the
Cone() constructor again.