src/sage/symbolic/integration: better giac errors
If the optional sagemath-giac is not installed, requesting integration
via giac "does nothing," i.e.

  sage: integrate(sin(x), x, algorithm="giac")
  integrate(sin(x), x)

This is due mainly to the fact that, when no algorithm is requested,
ALL available integration routines are tried one after the
other. Having the giac integrator do nothing so that the problem can
be handed to the next integrator makes sense.

However, when algorithm="giac" is explicitly requested, this is
confusing. This commit adds "giac" and "libgiac" to the chain of
integration backends only when sagemath-giac is present. The end
result is that passing algorithm="giac" to integrate() or integral()
will raise an error about the missing feature rather than quietly
doing nothing with the input:

  FeatureNotPresentError: sage.libs.giac is not available.
  Optional backend for giac integration not installed.
  No equivalent system packages for gentoo are known to Sage.