src/sage/groups/perm_gps/permgroup.py: bullet-proof a test
There's a test in this file that finds a random subgroup of
SymmetricGroup(6) with three generators, round-trips it through a
finite presentation, and then checks the result for isomorphism with
itself. If you are unlucky, however, GAP will get confused, and raise
a GAPError during the round trip. This was originally reported in
issue 32141, and a workaround was added at some point.

The workaround is a bit ugly though, and it does not make a lot of
sense to test something that in fact does not work. I have tried all
2024 of them, and the same problem does not arise with subgroups of
SymmetricGroup(4), so this commit reverts the workaround and then
changes the 6 to a 4. (Five has the same problem that six does.)

Closes: https://github.com/sagemath/sage/issues/32141