src/sage/geometry/cone.py: use face lattice to speed up isomorphisms
Rather than check all maps sending (a full-dimensional set of) extreme
rays of one cone to another for the isomorphism property, we need only
check those that are isomorphisms of the cones' face lattices. This
requires us to construct a Cone() from each cone's pointed-solid
component, and then take a tour through graph theory while fudging the
distinction between a 1d cone and the ray that generates it, but the
proof is in the pudding: the resulting algorithm is much faster.