src/sage/matrix/matrix2.pyx: rework some "extend" doctests
One of the doctests for solving a linear system without extending the
base ring is taking a long time,

  File "src/sage/matrix/matrix2.pyx", line 906, in
  sage.matrix.matrix2.Matrix.solve_right
  Warning: slow doctest:
    v = m.solve_right(vector(ZZ, [2]*500), extend=False)
  Test ran for 29.17s cpu, 29.18s wall
  Check ran for 0.00s cpu, 0.00s wall

This and the surrounding tests are solving large systems for no
apparent reason; we can test extend=False on a smaller system. This
commit rewrites these tests to solve systems at most an order of
magnitude smaller.