src/sage/crypto/block_cipher/des.py: run fewer DES test cases
We have one DES doctest that checks all of the test vectors given in
its reference, but checking them is slow:

  2025-07-24T02:20:21.8128567Z ##[warning]slow doctest:
  2025-07-24T02:20:21.8129623Z     for K, P, C in test: # long time
  ...
  2025-07-24T02:20:21.8131458Z Test ran for 31.91s cpu, 31.97s wall
  2025-07-24T02:20:21.8131872Z Check ran for 0.00s cpu, 0.00s wall

Rather than check them all (there are 171 in total), we may check only
a random subset. This commit takes a random sample of 5 test vectors,
and checks those. This speeds up the test significantly and should
still detect problems.