meson: make several optional features explicitly configurable
Several libraries are detected at setup time, and the corresponding
feature is enabled or disabled based on the library's presence. This
commit retains this behavior by default while making it possible to
explicitly enable or disable the feature via,

  $ meson setup -Dfoo=disabled

or

  $ meson setup -Dfoo=enabled

The former disables the feature even if the library is installed, and the
latter makes it an error if the library is not installed.

The list of features (valid values for foo) is,

  * bliss
  * brial
  * coxeter3
  * mcqd
  * meataxe
  * rankwidth
  * sirocco
  * tdlib