src/sage/structure/category_object.pyx: make variable_names() consistent
The variable_names() method is documented to return a tuple of
strings, but if there are no variable names it currently raises an
error. This is inconsistent with the documentation, but maybe more
importantly, inconsistent with the implementation in some rings like
RR where an empty tuple is returned.

This commit changes the "no variables" result to the empty tuple (),
with the justification that it's much easier to loop through or check
for an empty tuple than it is an either-a-tuple-or-an-error.