src/sage/groups/finitely_presented.py: variable_names() fallout
One test in this file is trying to inject the (nonexistent) variable
names from the finitely-presented group A5, and is expecting an
error. The implementation of variable_names() has been changed,
however, and now returns an empty tuple rather than an error when
there are no variables. As a result, A5.inject_variables() injects no
variables into the session, i.e. does nothing.

It seems like a weird thing to be testing in the first place, so this
commit removes the test rather than updating the expected output.

In another spot, the call to variable_names() is trying to catch the
ValueError that we no longer throw. This has been changed to an "if",
to catch the empty tuple instead.