src/sage/rings/qqbar.py: drop AA.__new__() method
This method was added to support unpickling of older non-singleton
versions of the AA object, but the test for it happens to be loading
such an old version of AA into scope, making it difficult to doctest
newly-fixed bugs. Since this class has been a singleton since commit
dfbb8bba in 2014, I think it is time to lay this method and its test
to rest.