src/sage/features/eclib.py: new feature for the eclib package
Our meson build system is capable of building sagelib without the
sage.libs.eclib.mwrank module requiring libec from the "eclib"
package. Here we add a new feature to represent this module. This will
allow us to use "needs eclib" in the tests that require it, leading
eventually to a passing test suite on no-eclib builds of sage.