src/sage/libs/eclib: needs sage.libs.eclib
Sage can now be built without the sage.libs.eclib.* extension modules,
but when it is, all of the tests under sage.libs.eclib fail. Here we
add "needs sage.libs.eclib" to the header of every (source code) file
under sage.libs.eclib, to indicate that those tests should be skipped
if the corresponding module was not built.