src/sage/interfaces/maxima_lib.py: add back two abs_integrate examples
Replace two examples that can only be solved when abs_integrate is
loaded. These examples are now tested to NOT produce a meaningful
result, so that if in the future they can be solved without
abs_integrate, we will know. The surrounding text explains why
the abs_integrate package is not loaded by default, and how to
load it manually.