src/sage/doctest/util.py: handle Ctrl-C from GAP
We've found that GAP's own SIGINT handler is less crashy than if we
mix GAP_Enter/GAP_Leave with cysignals' sig_on and sig_off. We've
installed that same handler for SIGALRM, but the code that catches it
can't tell the difference and converts them both in to
KeyboardInterrupt. To retain the ability to doctest this stuff, we have
to catch those KeyboardInterrupts and poke at them to see if they arose
from GAP.